Création de 2 hooks et de plusieurs components pour simplififier le code du composant SelectedIngredients
